Not yet working, but close
At the moment this code iterates over the content. Where it identifies a big role title, it combs through and pulls out that title and the associated skills. The next step is to identify the role- and skill-levels, and pull those out too. Then there's a small piece of work to be done to throw everything back out into a CSV.